Move code into HTTPUtil to make it more generically usable.
